pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-trust-gateway"
version = "0.1.0"
description = "Trust, reputation, and economic accountability layer for MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "A2A Settlement" }]
keywords = ["mcp", "trust", "oauth", "a2a", "settlement", "reputation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "PyJWT>=2.8.0",
    "starlette>=0.36.0",
    "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/a2a-settlement/mcp-trust-gateway"
Repository = "https://github.com/a2a-settlement/mcp-trust-gateway"
Specification = "https://github.com/a2a-settlement/mcp-trust-gateway/blob/main/SPEC.md"

[project.scripts]
mcp-trust-gateway = "mcp_trust_gateway.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_trust_gateway"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
